pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dailyresearchfeeder"
version = "0.0.1"
description = "Configurable daily research digest generator for papers and news feeds."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Yif Yang" }
]
keywords = ["research", "digest", "arxiv", "rss", "llm", "automation"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
]
dependencies = [
  "aiohttp>=3.9.0",
  "feedparser>=6.0.11",
  "openai>=1.70.0",
  "PyYAML>=6.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.2",
  "pytest>=8.3.0",
  "twine>=5.1.1",
]

[project.scripts]
dailyresearchfeeder = "dailyresearchfeeder.cli:main"

[project.urls]
Homepage = "https://github.com/Yif-Yang/DailyResearchFeeder"
Repository = "https://github.com/Yif-Yang/DailyResearchFeeder"
Issues = "https://github.com/Yif-Yang/DailyResearchFeeder/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["dailyresearchfeeder*"]

[tool.setuptools.package-data]
dailyresearchfeeder = [
  "templates/config.yaml",
  "templates/.env.example",
  "templates/user/*.txt",
  "templates/user/*.yaml",
]

[tool.setuptools.exclude-package-data]
"*" = ["__pycache__/*", "*.py[cod]"]
